requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
apipkg==1.5
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
enum34==1.1.6
et-xmlfile==1.0.1
execnet==1.7.1
glob2==0.7
greenlet==0.4.15
importlib-metadata==0.23
jdcal==1.4.1
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
namedlist==1.7
numpy==1.17.2
openpyxl==3.0.0
packaging==19.2
pandas==0.25.1
parse==1.12.1
parse-type==0.5.2
pluggy==0.13.0
py==1.8.0
pyodbc==4.0.27
pyparsing==2.4.2
pytest==5.1.3
pytest-assume==2.2.0
pytest-bdd==3.2.1
pytest-forked==1.0.2
pytest-html==2.0.0
pytest-metadata==1.8.0
pytest-parallel==0.0.9
pytest-xdist==1.29.0
python-dateutil==2.8.0
pytz==2019.3
pyzmq==18.1.0
selenium==3.141.0
six==1.12.0
urllib3==1.25.6
wcwidth==0.1.7
zipp==0.6.0
